Extract function `kompile_target`
